requirements.txt
langchain-openai>=0.0.5
openai>=1.0.0
python-dotenv>=1.0.0
fastapi>=0.100.0
pydantic>=2.0.0
yfinance>=0.2.0
langchain>=0.1.0 